feat(placement): dynamic panel sizing + category-based split ratios
Components no longer overflow/clip when placed in a smaller grid cell, and
multi-panel scenes are no longer a rigid 50/50 split.

Frontend:
- PanelCell wraps each component in a fit-to-width/fill-height scale-to-fit box
  (design canvas 1920 → scaled to the deterministically-computed cell px), so
  fixed-pixel components adapt to any cell with no letterbox; full-width panels
  keep scale=1 (no regression).
- New layout.ts: pure fraction/size math. Column fractions honor explicit
  size_ratio, else infer each area's width from its component category
  (diagram/algorithm 1.8, chart/timeline 1.3, text/list 1) so a diagram beside
  text auto-splits ~1.8:1 instead of 50/50.
- New PanelSizeContext/usePanelSize; RecursionTree/GraphTraversal/AnimatedTitle
  now size to the panel's design box instead of the full composition.

Backend Director:
- _normalize_plan infers size_ratio from component category when the LLM omits
  it; size_ratio widened to float across schemas; prompt rule 4 clarified.

Verified: layout self-check (scripts/check-layout.ts), tsc clean, backend
_normalize_plan inference check, and before/after stills (overflowing diagram
now contained; full-width scenes unchanged).

+++ backend/agents/director/schema.py
@@ -23,7 +23,7 @@ class PanelBlueprint(BaseModel):
 
     area: str    # "panel" | "left" | "right" | "main" | "sidebar"
     type: str = Field(description="The component ID (e.g. 'BulletList', 'ArchitectureDiagram')")
-    size_ratio: int = Field(default=1, description="Relative width proportion for grid layouts")
+    size_ratio: float = Field(default=1.0, description="Relative width proportion for grid layouts")
     delay_frames: int = Field(default=0, description="Delay in frames before this component appears in the scene")
     # data is omitted; that is populated by Scriptwriter/VisualArchitect
