+++ backend/models/video_script.py
@@ -50,7 +50,7 @@ class _NullSafeBase(BaseModel):
 class Panel(BaseModel):
     area: str   # "panel" | "left" | "right" | "main" | "sidebar"
     type: str   # one of the 13 component names
-    size_ratio: int = 1
+    size_ratio: float = 1.0
     delay_frames: Optional[int] = None
     data: dict[str, Any] = Field(default_factory=dict)
